6-size.c

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
-	printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long: %lu byte(s)\n", (unsigned long)sizeof(long long int));
-	printf("Size of a float: %lu byte(S)\n", (unsigned long)sizeof(float));
-	return (0);
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(int));
+printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+printf("Size of a long long: %lu byte(s)\n", sizeof(long long int));
+printf("Size of a float: %lu byte(S)\n", (unsigned long)sizeof(float));
+return (0);
 }
